refactor: replace old page with new component
