+++ pages/snek/create.vue
@@ -5,6 +5,7 @@
         <component
           :is="x"
           :show-explainer-text="false"
+          :class-column="false"
           @navigateToCreateNftTab="switchToNft" />
       </NeoTabItem>
     </NeoTabs>
@@ -14,8 +15,7 @@
 <script lang="ts">
 import { NeoTabItem, NeoTabs } from '@kodadot1/brick'
 
-const Collection = () =>
-  import('@/components/bsx/Create/CreateCollectionPage.vue')
+const Collection = () => import('@/components/create/CreateCollection.vue')
 const NFT = () => import('@/components/bsx/Create/CreateToken.vue')
 
 const components = { Collection, NFT, NeoTabItem, NeoTabs }
